vista de consejeros, validar si tiene una sesion o varias, menu administrativo tambien, slide anuncios

+++ app/Http/Controllers/Admin/InscripcioneController.php
@@ -213,8 +213,10 @@ class InscripcioneController extends Controller
         $personale = $inscripcione->personale; 
         $user = $personale->user;
         $inscripcione->delete();
-        $personale->delete();
-        $user->delete();
+        if(!count($personale->inscripciones)){
+            $personale->delete();
+            $user->delete();
+        }
 
         return redirect()->route('admin.inscripciones.index')->with('eliminar','Ok');
     }
